# perf: replace git-apply-delta with hand-rolled code