--- a/package.json
+++ b/package.json
@@ -55,7 +55,6 @@
     "clean-git-ref": "^2.0.1",
     "crc-32": "^1.2.0",
     "diff3": "0.0.3",
-    "git-apply-delta": "0.0.7",
     "ignore": "^5.1.4",
     "minimisted": "^2.0.0",
     "pako": "^1.0.10",