--- a/src/models/GitPackIndex.js
+++ b/src/models/GitPackIndex.js
@@ -1,13 +1,12 @@
 import crc32 from 'crc-32'
-import applyDelta from 'git-apply-delta'
 
 import { InternalError } from '../errors/InternalError.js'
+import { GitObject } from '../models/GitObject'
 import { BufferCursor } from '../utils/BufferCursor.js'
+import { applyDelta } from '../utils/applyDelta.js'
 import { listpack } from '../utils/git-list-pack.js'
 import { inflate } from '../utils/inflate.js'
 import { shasum } from '../utils/shasum.js'
-
-import { GitObject } from './GitObject'
 
 function decodeVarInt(reader) {
   const bytes = []