Replaces the `git-apply-delta` dependency with my own code. Shaves a little over a 1KB off the bundle gzip size (68.73KB -> 67.59KB), and eliminates the following direct and indirect dependencies:

- git-apply-delta@0.0.7
- varint@0.0.3
- bops@0.0.7
- to-utf8@0.0.1
- base64-js@0.0.2

Results are mixed as to whether my version is _faster_ but the speed is negligible anyway. 

cloning https://github.com/git/git.git
my git-apply-delta: 10s
old git-apply-delta: 11s
total time: 4 minutes

cloning https://gitlab.com/gitlab-org/gitlab.git
my git-apply-delta: 87s
old git-apply-delta: 79s
total time: 25 minutes


The _real_ reason I did this, is I'm considering implementing xdelta compression so we can generate efficient sized packfiles (which I consider a prerequisite to adding garbage collection), and rather than dive straight into compression, I thought I'd implement decompression first to make sure I know what I'm doing.